requirements.txt
pypdf>=5.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
openpyxl>=3.1.0
pymupdf>=1.25.0
